requirements.txt
flask==2.3.2
aiobotocore==2.5.0
requests==2.31.0
python-dotenv==0.21.0